requirements.txt
streamlit==1.35.0
cohere==4.40
python-pptx==0.6.23
python-docx==1.1.0
PyMuPDF==1.23.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
markdown==3.6
faiss-cpu
